refactor review recommendation algos to services package
